Implement persistent storage using the file system
Implements persistent storage for Raft using the filesystem directly.
The structure of the data is simple enough that it can be easily mapped
directly to the file system. Additionally, the Raft engine runs in a
single thread and so locks do not need to be included to guard against
race conditions.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>