Remove expect when stepping raw_node
Removes the `expect` that was used when stepping the raw_node in
`on_peer_message` and instead log any errors. raw_node returned an error
when a message was received from an unknown peer (i.e. when a node's
membership was removed but it still sent messages to the leader) which
cause the engine to crash.

Signed-off-by: Logan Seeley <seeley@bitwise.io>