Add applied() and set_applied() to StorageExt
Adds the `applied()` and `set_applied()` methods to the StorageExt trait
and the storage implementations. These methods get and set
(respectively) the index of the Raft entry that corresponds to the last
block that was committed on that node. This will be used inform the Raft
library of which entries it needs to send to the engine when the node
restarts
(https://docs.rs/raft/0.3.1/raft/struct.Config.html#structfield.applied).

Signed-off-by: Logan Seeley <seeley@bitwise.io>