Allow limiting number of transactions in a batch for workload
Workload command / docker compose file is used to perform LR tests.
There could be a case where test is required to be run with
n number of transactions put in a batch. This is similar to the
configuration option there is already for batch submission rate.
Default value being set to 1.

Signed-off-by: S m, Aruna <aruna.s.m@intel.com>